Moved most of Profiles dialog to a separate file
